requirements.txt
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.10
django-debug-toolbar==1.9.1
django-heroku==0.3.1
django-rest-auth==0.9.3
django-webpack-loader==0.6.0
djangorestframework==3.9.4
gunicorn==19.9.0
idna==2.6
psycopg2==2.8.2
pytz==2018.3
requests==2.22.0
six==1.11.0
sqlparse==0.2.4
urllib3==1.24.2
whitenoise==4.1.2
